requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.1

# Data Processing
pandas==2.1.4
numpy==1.24.4

# Async & HTTP Client
httpx==0.25.2

# Task Queue (for future Kafka/Celery integration)
celery==5.3.4
redis==5.0.1

# Message Queue (for future Kafka integration)
kafka-python==2.0.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Network Graph (for data generator)
networkx>=3.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0


requests==2.31.0